split tab Ui to individual files
