+++ video_blender_projects.csv
@@ -1,2 +1,6 @@
 project_name,project_path,frames1_path,frames2_path
-"Untitled",,,
+Untitled,,,
+baby jerry,C:\CONTENT\VFIformer-WebUI EXAMPLES\BLENDER\BABY JERRY AND AUNTS CC CORRECTED,C:\CONTENT\VFIformer-WebUI EXAMPLES\BLENDER\BABY JERRY AND AUNTS CC RAW,C:\CONTENT\VFIformer-WebUI EXAMPLES\BLENDER\BABY JERRY AND AUNTS CC RESYNTH
+BLACK SWEAT,C:\CONTENT\BLACK WORKOUT2-\SCENE2\blender,C:\CONTENT\BLACK WORKOUT2-\SCENE2\source,C:\CONTENT\BLACK WORKOUT2-\SCENE2\evensodds
+The Boo Boo Hogsett,C:\CONTENT\BABY PICTURES\THE BOO BOO HOGSETT\RESTORATION,C:\CONTENT\BABY PICTURES\THE BOO BOO HOGSETT\RAW SOURCE,C:\CONTENT\BABY PICTURES\THE BOO BOO HOGSETT\RESYNTH
+Boo Boo Hogsett,C:\CONTENT\VFIformer-WebUI EXAMPLES\BLENDER\BABY JERRY AND AUNTS CC CORRECTED,C:\CONTENT\VFIformer-WebUI EXAMPLES\BLENDER\BABY JERRY AND AUNTS CC RAW,C:\CONTENT\VFIformer-WebUI EXAMPLES\BLENDER\BABY JERRY AND AUNTS CC RESYNTH
